+++ NEWS.md
@@ -1,12 +1,33 @@
+- v0.0.4 (upcoming)
+ - Support `fma` for `Quantity`s.
+ - Promotion will wrap usual `Number` types in dimensionless, unitless `Quantity`
+   types. With `Quantity`s it is not always possible to promote to a common
+   concrete type, but we can at least ensure that the numeric backing types are
+   all promoted: (`promote(1.0u"m", 1u"N"//2, 0x08) == (1.0 m,0.5 N,8.0)`).
+   The usual outer constructor for `Quantity`s (`Quantity(val::T, unit)`)
+   continues to return a number of type `T` if the unit is `NoUnits`,
+   since most of the time the user does not want a dimensionless, unitless
+   quantity.
+
 - v0.0.3
+ - Bug fix: `uconvert(°C, 0x01°C)` no longer disturbs the numeric type
+ - Allow μ-prefixed units to be typed with option-m on a Mac, in addition to
+   using Unicode. Previously only `μm` could be typed this way.
  - Include a `baremodule` called `SIUnits` in the factory defaults. You can
    now do `using Unitful.SIUnits` to bring all of the SI units into the calling
    namespace.
- - Added units to the factory defaults: `sr` (steradian), `lm` (luminous flux),
-   `lx` (illuminance), `Bq` (becquerel), `Gy` (gray), `Sv` (sievert), `kat` (katal).
- - Simplify array creation, as in `[1, 2]u"km"` (#29)
- - Support multiplying ranges by units, as in `(1:3)*mm` (#28)
- - Bug fix (#26)
+ - Added remaining SI units to the factory defaults: `sr` (steradian), `lm`
+   (luminous flux), `lx` (illuminance), `Bq` (becquerel), `Gy` (gray),
+   `Sv` (sievert), `kat` (katal).
+ - Simplify array creation, as in `[1, 2]u"km"` [#29](https://github.com/ajkeller34/Unitful.jl/pull/29)
+ - Support multiplying ranges by units, as in `(1:3)*mm` [#28](https://github.com/ajkeller34/Unitful.jl/pull/28)
+ - Bug fix [#26](https://github.com/ajkeller34/Unitful.jl/issues/26)
+ - Promoting `Quantity`s with different dimensions now returns quantities with
+   the same numeric backing type, e.g. `Quantity{Float64}`. Ideally, this would
+   also be true if you mixed unitless and unitful numbers during promotion, but
+   that is not yet the case. See [#24](https://github.com/ajkeller34/Unitful.jl/issues/24)
+   for motivation.
+
 
 - v0.0.2
  - Bug fixes (`[1.0m, 2.0m] ./ 3` would throw a `Unitful.DimensionError()`).